添加课程删除功能

+++ service/service-edu/src/main/java/com/atguigu/serviceedu/service/impl/EduVideoServiceImpl.java
@@ -6,11 +6,16 @@ import com.atguigu.serviceedu.client.VodClient;
 import com.atguigu.serviceedu.entity.EduVideo;
 import com.atguigu.serviceedu.mapper.EduVideoMapper;
 import com.atguigu.serviceedu.service.EduVideoService;
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
+
+import java.net.SocketTimeoutException;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * <p>
@@ -23,17 +28,63 @@ import org.springframework.web.multipart.MultipartFile;
 @Service
 public class EduVideoServiceImpl extends ServiceImpl<EduVideoMapper, EduVideo> implements EduVideoService {
 
+
     @Autowired
     private VodClient vodClient;
 
     @Override
-    public void removeVideoById(String videoId) {
-        EduVideo video = baseMapper.selectById(videoId);
-        String videoSourceId = video.getVideoSourceId();
+    public void removeVideoByCourseId(String courseId) {
+        QueryWrapper<EduVideo> queryWrapper = new QueryWrapper<>();
+
+        queryWrapper.eq("course_id", courseId);
+        List<EduVideo> videos = baseMapper.selectList(queryWrapper);
+
+        //获取视频id
+        ArrayList<String> videoSourceIdList = new ArrayList<>();
+        for (EduVideo video : videos) {
+            String videoSourceId = video.getVideoSourceId();
+            if (!StringUtils.isEmpty(videoSourceId)) {
+                videoSourceIdList.add(videoSourceId);
+            }
+        }
+        // 如果存在视频就批量删除
+        if (videoSourceIdList.size() > 0) {
+            String videoSourceIdsString = StringUtils.join(videoSourceIdList.toArray(), ",");
+            vodClient.removeVideos(videoSourceIdsString);
+        }
+        //删除小节本身
+        baseMapper.delete(queryWrapper);
+
+    }
+
+    @Override
+    public void removeVideoSource(String videoSourceId) {
+        vodClient.removeVideos(videoSourceId);
+    }
+
+    @Override
+    public void removeVideo(EduVideo eduVideo) {
+        String videoId = eduVideo.getId();
+        String videoSourceId = eduVideo.getVideoSourceId();
+
+        // 如果没有阿里云视频id，并且有小节id，那就在数据库里面查一下有没有阿里云视频id
+        if (StringUtils.isEmpty(videoSourceId) && !StringUtils.isEmpty(videoId)) {
+            System.out.println("if videoId" + videoId);
+            EduVideo videoFormDb = baseMapper.selectById(videoId);
+            if (!StringUtils.isEmpty(videoFormDb.getVideoSourceId())) {
+                videoSourceId = videoFormDb.getVideoSourceId();
+                System.out.println("videoSourceId " + videoSourceId);
+            }
+        }
+        //删除视频id
         if (!StringUtils.isEmpty(videoSourceId)) {
+            System.out.println("del videoSourceId " + videoSourceId);
             vodClient.removeVideos(videoSourceId);
         }
-        baseMapper.deleteById(videoId);
+        // 删除小节
+        if (!StringUtils.isEmpty(videoId)) {
+            baseMapper.deleteById(videoId);
+        }
     }
 
     @Override
